refactor: 잘못작성한 패키지명 수정
- scehdule -> schedule 변경
