+++ src/main/java/com/dmforu/schedule/ScheduleController.java
@@ -1,4 +1,4 @@
-package com.dmforu.scehdule;
+package com.dmforu.schedule;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
